Creamos la clase laboratorio,separando la definicion(.h) de la implementación(.cpp). Consta de 2 atributos: un contador y un arreglo de tipo computadora; además de que cuenta con su constructor base, y los métodos agregarFinal() y mostrar(). Finalmente, instanciamos un objeto de esta clase, y mediante su método agregarFinal() anexamos los objetos de tipo computadora creados previamente, y luego usamos el método mostrar() para imprimir en pantalla los objetos creados.

+++ computadora.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 Computadora::Computadora() {}
 
-Computadora::Computadora(const std::string &so, const std::string &cpu, int ram, float almacenamiento)
+Computadora::Computadora(const std::string &so, const std::string &cpu, float ram, float almacenamiento)
 {
     this->so = so;
     this->cpu = cpu;
@@ -28,7 +28,7 @@ std::string Computadora::getCPU()
 {
     return cpu;
 }
-void Computadora::setRAM(int ram)
+void Computadora::setRAM(float ram)
 {
     this->ram = ram;
 }
